Use correct size type so that things are 64-bit clean.
git-svn-id: https://svn.salilab.org/imp/trunk@429 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ kernel/src/unary_functions/OpenCubicSpline.cpp
@@ -51,10 +51,10 @@ Float OpenCubicSpline::operator()(Float feature)
   }
 
   // determine bin index and thus the cubic fragment to use:
-  unsigned lowbin = static_cast<int>((feature - minrange_) / spacing_);
+  size_t lowbin = static_cast<size_t>((feature - minrange_) / spacing_);
   // handle the case where feature ~= maxrange
   lowbin = std::min(lowbin, values_.size() - 2);
-  unsigned highbin = lowbin + 1;
+  size_t highbin = lowbin + 1;
   Float lowfeature = minrange_ + lowbin * spacing_;
 
   Float b = (feature - lowfeature) / spacing_;
@@ -68,10 +68,10 @@ Float OpenCubicSpline::operator()(Float feature)
 
 Float OpenCubicSpline::operator()(Float feature, Float& deriv)
 {
-  unsigned lowbin = static_cast<int>((feature - minrange_) / spacing_);
+  size_t lowbin = static_cast<size_t>((feature - minrange_) / spacing_);
   // handle the case where feature ~= maxrange
   lowbin = std::min(lowbin, values_.size() - 2);
-  unsigned highbin = lowbin + 1;
+  size_t highbin = lowbin + 1;
   Float lowfeature = minrange_ + lowbin * spacing_;
 
   Float b = (feature - lowfeature) / spacing_;
